se implemento el useHandlePhotoChange, para separar esa logica que podria ser reutilizable y se hicieron los ajustes en el perfil
